fix: refactored and removed unused files
