--- src/common/parsers/schemaParser.ts
@@ -1,11 +1,0 @@
-import Ajv from "ajv";
-import { DatabaseSchemaSchema } from "../ajvSchemas";
-
-// To handle recursive schema, you would typically set it up in your AJV instance like this:
-export const parse = (data: any) => {
-  const ajv = new Ajv({ allowUnionTypes: true });
-  ajv.addSchema(DatabaseSchemaSchema, "DatabaseSchema");
-  DatabaseSchemaSchema.properties.externalSchemas.additionalProperties = {
-    $ref: "DatabaseSchemaSchema",
-  };
-};
